# api/uspace/uspacedocs_swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for submitting/monitoring jobs to/from an execution machine",
        "title": "Uspace API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8079",
    "basePath": "/api/v1",
    "paths": {
...
        "/job": {
            "get": {
                "description": "GET retrieves jobs by uid(s), jid, or returns all. POST submits one or multiple jobs.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get or submit jobs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Limit number of jobs",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated list of user IDs",
                        "name": "uids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Job ID or '*' for all jobs",
                        "name": "jids",
                        "in": "query"
                    },
                    {
                        "description": "Single job",
                        "name": "job",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/kyri56xcaesar_kuspace_internal_utils.Job"
                        }
                    },
                    {
                        "description": "Multiple jobs",
                        "name": "jobs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/kyri56xcaesar_kuspace_internal_utils.Job"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "GET retrieves jobs by uid(s), jid, or returns all. POST submits one or multiple jobs.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get or submit jobs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Limit number of jobs",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated list of user IDs",
                        "name": "uids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Job ID or '*' for all jobs",
                        "name": "jids",
                        "in": "query"
                    },
                    {
                        "description": "Single job",
                        "name": "job",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/kyri56xcaesar_kuspace_internal_utils.Job"
                        }
                    },
                    {
                        "description": "Multiple jobs",
                        "name": "jobs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/kyri56xcaesar_kuspace_internal_utils.Job"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "kyri56xcaesar_kuspace_internal_utils.Job": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "boolean"
                },
                "completed_at": {
                    "type": "string"
                },
                "cpu_limit": {
                    "type": "string"
                },
                "cpu_request": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "number"
                },
                "env": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "ephimeral_storage_limit": {
                    "type": "string"
                },
                "ephimeral_storage_request": {
                    "type": "string"
                },
                "input": {
                    "type": "string"
                },
                "input_format": {
                    "description": "perhaps unecessary",
                    "type": "string"
                },
                "jid": {
                    "type": "integer"
                },
                "logic": {
                    "type": "string"
                },
                "logic_body": {
                    "type": "string"
                },
                "logic_headers": {
                    "type": "string"
                },
                "memory_limit": {
                    "type": "string"
                },
                "memory_request": {
                    "type": "string"
                },
                "output": {
                    "type": "string"
                },
                "output_format": {
                    "type": "string"
                },
                "parallelism": {
                    "type": "integer"
                },
                "params": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "priority": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "timeout": {
                    "description": "in minutes",
                    "type": "integer"
                },
                "uid": {
                    "type": "integer"
                }
            }
        },
...
    }
}